refactor: simplify dependency structure on menu
